syscalls: implement internal callbacks for each system call
The included python script parses function declarations and uses
the ARM or x86 system call ABI to autogenerate declarations and
empty definitions of callbacks taking the arguments of each system
call.
The default definition is weakly linked, so the user can override it.
The included system call definitions and weak_callbacks.* are valid
for ARM Linux.