taint: use LabelSet struct in cleaner more C++ish way
First, call LabelSet's constructor and destructor. We use the "placement new"
operator to allow us to keep tracking its allocation in the mymem pool.
Second, repace the Bitset* member with a Bitset, since it can never be null
or shared and we dereference it for every operation. Memory management of the
Bitset object is now handled automatically by LabelSet's ctor an dtor.