syscalls: add PPP callbacks for sysrets
Replace the default weakly-linked syscall handlers with weakly-linked handlers
that check if there are any PPP callbacks for the return point of the syscall.
If there are, they save all the syscall's arguments to a struct which will be
passed to a function that calls the PPP callbacks at the return point.

The default callbacks are #included into syscalls.cpp because it uses PPP
macros, which generate and access arrays, and there's no way yet to generate
extern definitions of them in other files (ie. in default_callbacks.cpp)

If any of these default handlers is overridden, the PPP sysret callback for
that syscall will no longer be called. Since the PPP structures are only
accessible from syscalls.cpp, and the autogenerated callbacks already exist in
syscalls.cpp, an override that preserves PPP functionality must be done by
manually changing the autogerated callback definitions.
This commit comments out override callbacks that existed in sycalls.cpp